infrastructure/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.26.0
requests>=2.28.0
